requirements.txt
openai
torch
wandb
datasets
peft
trl 
transformers
tensorflow
google-generativeai
chardet
